Complete overhaul of tagging module.
Now using experimental optional attribute, it now takes in just 3
variables (tags, extra_tags and additional_tag_maps)

The tags can now be passed into a map and the keys it understand it will
convert properly into pascal case and present out along with it being
presented as a list of maps for resources which require that format.

also re-organised to match other module layouts.

	new file:   data.tf
	new file:   locals.tf
	modified:   main.tf
	modified:   outputs.tf
	modified:   variables.tf